add-cluster: user to review images at key points
Extract a new function '_user_review_image' and use it to allow the user
to review the generated images at key points.

The less steps that require review, the faster the workflow will be so
only final results are reviewed.

Test Plan:
Try to add a new cluster with the 'mel add-cluster', fail with
NotImplementedError as expected.

On one run through, accept all results.
On another run through, reject the first result.