add-cluster: write images, not NotImplemented
Save the resulting ident images to the specified location, create the
location if it doesn't exist already.

No longer fail with NotImplemented, this is a usable implementation.

Add TODOs for some obvious refinements.

Test Plan:
Try to add a new cluster with the 'mel add-cluster', idents are written
as expected.