add-cluster: use date for filenames if possible
To make it easier to determine when ident images need a refresh, record
the date they were taken in the ident filename. Record the date as ISO
8601 for portability.

Try to determine the capture date from the source filename.

Test Plan:
Use 'mel add-cluster' to add a cluster from source photos with
filenames as generated by an iPhone. Target files are generated with ISO
8601 dates as expected.