Add 'mel micro-record' for making vids
Make a utility for recording videos of moles, this will make it much
easier to iterate on routines related to the identification of moles.

Just record a video straight out to begin with, in later work we'll add
parameters for controlling the output file names. Also, we'll
automatically identify when there is a mole on screen and start new
videos for each one.

Test Plan:
$ ./bin/mel micro-record
.. watch recorded video and see it's recorded successfully ..