micro-record: record multiple movies automatically
Try to record one mole per movie, automatically start and stop recording
by using the MoleAcquirer to tell us when we're looking at a new mole
and when we've stopped looking at a mole.

Add an argument for prefixing the movie files, so if we're recording
multiple movies of the same subject then it's easy to keep them
separated later.

Test Plan:
$ ./bin/mel micro-record --filename-prefix myprefix
.. record multiple movies, look ok ..