micro-replay: support vector machine to classify
Try to classify moles based on the stats we're extracting, use a
'support vector machine', which seems to be the method suggested by the
rather useful cheat sheet here:
http://scikit-learn.org/stable/tutorial/machine_learning_map/

With enough samples of each mole, we might be able to automatically
determine which is which. Extensive testing required.

Test Plan:
$ ./bin/mel micro-replay *.avi
.. successfully predicts first and last mole, so far so good ..