mel.lib.moleimaging: use aspect ratio, coverage%
Instead of using ellipse width and height as stats to identify moles,
use aspect ratio and relative size of the fitted ellipse to the mole.

These new values should vary independently of the area, unlike the width
and height. This makes them better candidates for classification with a
support vector machine.

Test Plan:
$ ./bin/mel micro-replay *.avi
$ ./bin/mel micro-analyse *.json
.. 100% accuracy ..

.. temporarily stop recording mole area ..
$ ./bin/mel micro-replay *.avi
$ ./bin/mel micro-analyse *.json
.. 90% accuracy, not bad for just those two ..