Mel: save image from vid replay, instead of stats
Make it less expensive to experiment with the calculation of the mole
stats by exporting images from the video replay instead of the stats
directly. Instead have 'mel micro-analyse' load images and calculate the
stats itself.

In later work we may add an extra utility to go from images -> stats, as
this analysis is not cheap.

Test Plan:
$ ./bin/mel micro-replay *.avi
$ ./bin/mel micro-analyse *.jpg
.. 100% accuracy ..