micro-add: show context, acquire mole, align, save
Require the user to provide the path of the mole to capture a microscope
image of.

Show the context image of the mole, so the user can verify the location
visually.

Use MoleAcquirer to determine that the stats can be consistantly measured.

Add a new function to overlay alignment hints on the incoming image.
Images which are normalised will hopefull be easier to match
mechinically and to compare manually.

Save the image when it is being consistantly measured and is aligned
within tolerances.

Test Plan:
Use micro-add to capture a couple of mole images.