micro-add: display all context images
Display constellation image when doing micro capture, this removes the need
to manually browse to it.

Display any image in or above the specified directory, this is so there
can be varying levels of context.

Only display images in or below the current directory, so the level of
context follows the current working directory.

Test Plan:
Capture a couple of constellation moles with micro-add.