find_mole_contour: find most central large contour
When looking a mole in a list of contours, return the most central large
contour. This has the effect of making it possible to 'select' a
particular mole by making it more central in the image.

Importantly we can't just select the most central contour, as this could
be a small artifact and not a mole.

Test Plan:
Use micro-add to select between adjacent moles, by moving the microscope
towards or away from the intended mole.