micro-add: show path in window names
Make it easier to determine the location of a mole to capture by putting
the path to the mole in the title of the capture windows. In particular
this makes it easier to tell which side of the body we're dealing with,
as images can be somewhat ambiguous.

Test Plan:
.. capture a few moles with add-micro ..