Remove import_all.py
Update static_tests.py to not run import_all.py.

Nose will automatically import all the modules anyway, during
unit_tests.sh. It does this in order to discover tests. No need to do
this twice.

Also, importing is more of a run-time test - it requires all the
run-time dependencies to be satisfied.