mel micro-add: single window to display images
Use the new mel.lib.ui.MultiImageDisplay to simplify the user interface
by requiring the user to interact with a single window, instead of
multiple windows.

Previously the multiple windows in the interface may have been
overlapping, which would require mouse work that would be unnecessary
with a single window interface.

Allow the user to supply the dimensions of the user interface as
arguments. In later work we will detect the dimensions of the monitor
and size appropriately.

Test Plan: Update a mole with 'mel micro-add'