micro-add: support '--min-compare-age-days'
Add an argument which we may use to control how old the comparison image
is. This can be useful where there are moles that have a long history
and look significantly different from the original image but are not
dangerous.

Test Plan:
'micro-add' with and without new argument, verify the expected
comparison images are displayed.