Add 'mel rotomap-view' command
Introduce the idea of a 'rotomap', where a series of images is taken
whilst a particular subject rotates clockwise, e.g. 'lower left leg'.

This provides excellent context reference for moles, as you can get a
very clear idea of where they are in relation to eachother when viewing
from multiple angles.

Provide a new command 'rotomap-view', which simply displays the images
in a 'rotomap'.

In later work we'll be able to record the locations of moles in the maps
so that they can be counted and tracked. This should provide a way to
monitor for new moles - by comparing which moles appear in old and new
rotomaps.

We will also have the possibility of automatically generating the
context images that are so useful for identifying moles. This should
work for both cluster images as well as single images.

Test Plan:
$ ./meta/run_tests.sh
.. view a simple rotomap with 'mel rotomap-view' ..