mel.lib.image: Help for centering image in region
Add calc_centering_offset() and translated_and_clipped(), which together
can create a version of an image which is clipped to a region and
centered about a certain point.

This can be useful for 'zooming in' on a particular point in a very
large image, which would need to be clipped to the confines of the
display.

Test Plan:
$ ./meta/run_tests.sh