rotomap-show: show multiple images, scale uuids
Take multiple images as parameters and show them with a blank line
in-between.

Adjust the number of characters of the uuid shown, so that moles appear
with unique identifiers.

Test Plan:
Successfully view all the images in a rotomap, try artificially
increasing the need for more uuid characters by choosing uuids that
collide.