rotomap-show: only minimum necessary digits
When it becomes necessary to show more than one digit to disambiguate
moles, only show the minimum necessary number of digits for each mole.

For example, taking three moles: [a1, a2, b1], it is only necessary to
show two digits for 'a1' and 'a2'. For 'b1' we need only show 'b', as no
other moles in the list start with 'b'.

Test Plan:
.. rotomap-show a series and verify that only the minimum is shown ..