rotomap-edit: only hex digits in uuids, add uuids
If a mole does not have a uuid then generate one to save with it.
Previously the mole would have to be in the old list form for this to
happen.

Use only hex digits in uuids, the dashes in the standard form don't
serve a purpose for us and may ultimately complicate things.

Test Plan:
.. Remove the uuid a mole and observe a new one is generated ..
.. Add a new mole and observe the uuid is just hex digits ..