rotomap-relate: relate moles better
If there are no moles in the target image then use a cut-off of zero
instead of None.

Given mole A in the source map and mole B in the target map, only add
mapping A -> B if A is also the closest match for B. This prevents the
situation where there are multiple possible matches for B and a worse
one is chosen.

Test Plan:
.. Successfully relate many maps ..