rotomap-edit: save moles when removing no mole
In the case where we try to remove a mole and none exists, still save
the list of moles anyway (which may be empty).

This allows us to save .json files for images that have no moles. It is
convenient to have .json files for every image, even if empty, for
consistency. This is especially useful when relating json files, so that
when there are empty files 'between' non-empty files, the non-empty
files are not erroneously related.

Test Plan:
.. shift-click on an empty image and observe a .json file is created ..
.. use right-arrow to view the next image ..
.. use left-arrow to view the original image, which loads ok ..