Add 'mel rotomap-uuid' for easier uuid handling
Just like as with Git hashes, the full uuid of moles can be unwieldy to
use on the command-line. Provide a way to obtain the full uuid from a
prefix in a number of rotomap json files.

This will make it easier to make e.g. bash autocomplete, or helper
scripts that can deal with a unique prefix as opposed to the whole uuid.

Of course, the more narrow the list of jsons, the more likely that a
particular prefix will be unique to that list.

Test Plan:

Check a prefix that doesn't exist

    $ mel rotomap-uuid notauuid *.json
    $ echo $?
    1

Check a prefix that matches only once

    $ mel rotomap-uuid 0 *.json
    08fb6d4bb4e54df9b1d2cc6c2848b667
    $ echo $?
    0

Check a prefix that matches more than once

    $ mel rotomap-uuid 4 *.json
    4797c5e1b1aa4ff7bb4f83169f9433d2
    4fd1749c95b94cdd8efc311a8b16dcba
    $ echo $?
    0