rotomapedit: base default display size on screen
Use the size of the screen as a guide when determining the default
display size.

The margins picked seem suitable on the particular machine it was tested
on, taking the display out of the way of the dock at the bottom of the
screen.. In later work we can try to better determine useful margin
sizes.