BREAK: rotomap-edit requires '--images' not PATH
Change the interface of 'mel rotomap-edit' to require an '--images'
parameter instead of a single path.

This allows for more flexibility around what is being edited, for
example it's now possible to edit images from different rotomaps
side-by-side.

This opens the possibility of being able to edit multiple rotomaps
side-by-side as well, by specifying '--images' multiple times. Perhaps
later we'll use the 'up' and 'down' arrow keys to allow quick movement
between rotomaps. This will be useful for marking up moles in a new
rotomap, using an old rotomap as reference.

Test Plan:
$ mel rotomap-edit --images *.jpg
.. seems to work well ..