BREAK: remove '--rot90' from rotomap-edit
Simplify rotomap-edit by removing the option to rotate images in
increments of 90 degrees.

This was originally added to help when working with images taken using
an iOS device, which will typically store the orientation of the device
in the exif data stored in the resulting images. Many programs will
honor the exif data and display the image with the correct rotation,
this does not happen when loading images with OpenCV.

It's simple enough to alter the image to be orientated correctly, by
using ImageMagick like so:

  $ mogrify -auto-orient image.jpg

In later work we should issue a warning if images are loaded that
have exif data with an alternate orientation.

Test Plan:
.. copy, paste, remove, move and add moles with rotomap-edit ..
.. zoom and unzoom ..
.. show and hide mole markers ..