Add 'mel-debug triangulate' to debug autopaste
Make it easier to see what's going on with
mel.rotomap.moles.get_best_moles_for_mapping() by adding a new
interactive command that visualizes the moles that are picked as the
best reference for a supplied point.

In later work we can hopefully refactor in such a way that duplication
is reduced between 'mel rotomap-edit' and this command.

Test Plan:
$ ./meta/run_tests.sh

.. Run 'mel-debug triangulate' on a series ..
    .. Verify that the expected points are highlighted ..