Add 'vulture' to static tests, report dead code
Code that is not used creates an extra burden when trying to test or
understand the code. It can be preserved in the history of the project
by removing it in a new commit.

Use the 'vulture' tool to assert that we don't have code that is unused.

Test Plan:
$ ./meta/run_tests.sh