FIX+BREAK: 'mel rotomap-molepicker' rot90, images
The '--rot90' and '--images' options were removed from rotomap-edit in
commits 3a74aa0 and 45bbbea respectively. The appropriate changes were
not made to 'mel rotomap-molepicker' at the same time.

Introduce the same breaking changes to rotomap-molepicker, which will at
the same time fix it.

'--rot90' was originally added to help when working with images taken
using an iOS device, which will typically store the orientation of the
device in the exif data stored in the resulting images. Many programs
will honor the exif data and display the image with the correct
rotation, this does not happen when loading images with OpenCV.

It's simple enough to alter the image to be orientated correctly, by
using ImageMagick like so:

    $ mogrify -auto-orient image.jpg

In later work we should issue a warning if images are loaded that
have exif data with an alternate orientation.

Requiring '--images' instead of 'PATH' allows for more flexibility
around what is being viewed.

This opens the possibility of being able to view multiple rotomaps
side-by-side as well, by specifying '--images' multiple times. Perhaps
later we'll use the 'up' and 'down' arrow keys to allow quick movement
between rotomaps.

Test Plan:
.. select a mole with rotomap-molepicker ..
.. go to previous and next images ..
.. zoom and unzoom ..
.. show and hide mole markers ..