Mel Rotomap: fade mole markers by default
Generally mole markers are useful for knowing where a mole has been
marked, having a rough idea that the uuids are distinct, and roughly
where they should be.

If the markers are somewhat faded then they seem to still
serve these purposes, and they also allow the mole underneath to be
seen. This makes it easier to verify precise placement over the mole and
also to check that it's the right mole.

Test Plan:
$ ./meta/run_tests.sh

.. mel rotomap-edit on an existing series ..
    .. observe markers faded by default ..
    .. flip fade and visibility on/off ..
    .. zoom in ..
    .. flip fade and visibility on/off ..