mel micro-add: better width and height defaults
If the user doesn't supply a width and height, use
mel.lib.ui.guess_fullscreen_width_height() to provide sensible defaults.

This appears to introduce the unwelcome side-effect of displaying an
empty window with the title 'tk'. It generally doesn't obscure anything
though, so a net positive perhaps.

Test Plan:
$ ./meta/run_tests.sh

.. Change to a mel data set ..
$ mel list --sort lastmicro | head -1 | xargs mel micro-add
.. Default size of window is good ..
.. 'tk' window is annoying, but not too annoying ..