mel rotomap-overview: search paths for rotomaps
This command is easier to use if it searches for rotomap directories for
itself, as opposed to requiring them to be supplied to it.

Now you can supply the parent directory of multiple rotomaps and it will
overview all of them.

Test Plan:
$ ./meta/run_tests.sh

.. In a rotomap data dir ..
$ mel rotomap-overview .
LeftArm/Lower/20160117T1757 ++++++++++++++++++++
LeftArm/Lower/20160117T1759 -------------------
RightArm/Lower/20160117T1802 +++++++++++++++++++++++++++++
RightArm/Lower/20160117T1803 +++++++++-+++++++++++