mel list: ignore all dirs starting with '__'
Don't just treat the '__micro__' directory as special, ignore all
sub-directories that start with '__' to provide better
forward-compatability and to allow arbitrary user-provided directories.

For example, it can be desirable to archive old context images under
'__arc__', so that they are preserved and discoverable but do not appear
when doing 'mel micro-add'.

Test Plan:
$ ./meta/run_tests.sh

In a mel data dir with a '__arc__' subdir:
$ mel list
.. the __arc__ subdir is not treated as a mole ..