Add mel rotomap-edit '--follow'
Introduce a rudimentary 'following' behaviour that allows you to
automatically centre and zoom the display on a particular mole. This is
intended to make it easier to verify that a particular mole is tagged
correctly and to make fine adjustments to it's position.

In later work, this should probably be controllable from the gui and not
just the initial command-line options.

Test Plan:
$ ./meta/run_tests.sh
.. 'mel rotomap-edit' --follow a mole around ..