mel rotomap-edit: move editor to front
Save a little bit of valuable time and remove frustration at keyboard
controls not working by bringing the displays to the front of the
screen. This gives keyboard focus and brings our windows in front of
other windows. May not work as expected if there are multiple Python
processes.

Test Plan:
$ ./meta/run_tests.sh
.. mel rotomap-edit to exiting map, keys work immediately ..