Fix mel rotomap-diff: print appropriate digits
Determine the appropriate number of digits to show before modifying the
uuids to start with '-', '+', etc. This prevents the situation where '-'
is determined to be unique enough to display.

This introduces the new problem of using more space than necessary in some
circumstances, which is more tolerable than the previous problem.

Test Plan:
$ ./meta/run_tests.sh
.. 'mel rotomap-diff' some files prev showing no digits ..