tidy: del unused params in mouse callback funcs
Our linters will complain about unused parameters. There seem to be a
number of these detected in the mouse callback functions.

After checking that there aren't any mistakenly unused, prefix the
unused parameters with an underscore. This will cause the linter to
ignore the unused parameter. Also, 'del' the unused parameters. This
will make it clearer to readers that they are not intended to be used.

Test Plan:
.. Upgrade linters ..
$ ./meta/run_tests.sh