find_mole_ellipse: abort if clipped
The previous behaviour was a bit confusing. The clipping was not
complete - it would only clip the top-left to the top-left. This was
ignoring three more clippings required, involving the bottom-right.

As this function is a 'best-effort' to find a mole ellipse, abort if any
clipping is required.

Introduce a new mel.lib.image funciton to help with this.

Test Plan:
$ ./meta/run_tests.sh
.. After error due to empty image slice during 'mel rotomap-edit' ..
    .. Apply patch ..
    .. Fail to repro ..