rotomap-display: distinct colouring, not from uuid
Colours were previously picked based on the uuid of moles. There are
inevitably co-incidences where the colours picked are the same or
similar.

It doesn't seem to be so useful for moles to have a consistently chosen
colour between runs of commands - they are not particularly memorable.

If we're free to choose arbitrary colours then instead we can choose
colours based on the limited number of moles available from the
arguments passed to rotomap-view and rotomap-edit.

Use the mel.rotomap.tricolour colouring scheme. So depending on the
number of moles to colour, moles can be coloured more distinctively.

Fixes #6.

Test Plan:
$ ./meta/run_tests.sh
.. Run mel rotomap-edit and observe that colours are better ..