mel micro-add: auto-rotate and centre images
Automatically rotate and centre mole images before saving them, to make
for easier comparisons later.

We could also apply the centring and rotation in the 'mel micro-compare'
tool, but this would mean using the special tool for easier comparisons.
If we bake it into the image then it will be visible everywhere.

This method is lossy, it doesn't seem like we need the border pixels for
anything.

Later we can try relaxing the position and rotation constraints on
images, whilst capturing good results. This should speed up collection.

Builds towards #3, #4.

Test Plan:
$ ./meta/run_tests.sh
.. Try capturing some moles, results are good ..