mel.rotomap.moles: add 'is_uuid_canonical' attrib
When loading moles, automatically add in an 'is_canonical' attribute.
This is used to determine whether moles have a uuid that has been
randomly generated and not matched to an existing mole, or if it is the
canonical uuid for that particular mole.

This will be used when automatically matching moles across images -
where moles have a uuid that is not canonical then it may be replaced.

Test Plan:
$ ./meta/run_tests.sh
.. Load and save some rotomaps with 'mel rotomap-edit' ..