Refactor, move MoleMarkerOverlay to Editor
In mel.rotomap.display, make maintaining the MoleMarkerOverlay a
responsibility of the Editor instead of the Display.

The Display doesn't need to know about moles, it only needs to be
supplied an overlay object.

This removes the need for some duplicated methods on the Display, it
also paves the way for some more complex overlays.

Test Plan:
$ ./meta/run_tests.sh
.. 'mel rotomap-edit', zoom around, follow moles, show/hide moles ..