rotomap-edit: press 'r' to auto-mark moles
Reduce time required to mark moles in an image by providing a reasonably
reliable 'auto-mark' function that uses OpenCV's SimpleBlobDetector to
look for likely moles in the image.

Moles will only be detected if they are part of the masked region of the
image. Experimentally it's been found that the SimpleBlobDetector will
unsurprisingly find lots of blobs elsewhere in images, resulting in
unmanageable amounts of false +ves.

Test Plan:
.. Automatically mark a whole rotomap, fixing errors ..
.. Observe that overall it seems to speed up the process ..