rotomap-edit: shift+rclick to re-init mole uuid
When auto-relating between rotomap images, it would be convenient to
have a 'partial undo'. This would be useful for removing spurious
matches.

Something that's almost as useful is being able to re-initialise the
uuid of a target mole, so that it is not canonical and doesn't match any
other uuid.

Another way to do this is to delete and recreate the mole, this has the
unwanted side-effect of losing the detected radius information.

The functionality for 'shift-lclick' is delete, so 'shift-rclick' was
picked for re-init as it's a little related.

Test Plan:
.. Auto-relate with 't' and undo mismatch ok with shift-rclick ..