rotomap-edit: alt+rclick to globally replace uuid
When auto-relating around a rotomap, there's a 'seam' when you wrap
around to the beginning again. This is because some moles in the first
image may be the same in the last image, but they will have acquired
different uuids.

Make it easier to deal with this seam by providing the ability to
globally remap a uuid to another.

The functionality for 'alt-lclick' is replace, so 'alt-rclick' was
chosen for global replace.

Test Plan:
.. Auto-relate with 't' and consolidate uuids with alt-rclick ..