rotomap-edit: more articulate mole-relate guesses
Provide better guesses when using 't' to auto-relate moles. Adjust the
guessed offset for a mole by taking into account the surrounding known
offsets individually. This is as opposed to the previous method of
using the average of all the offsets.

Weight the influence of the surrounding offsets by the inverse square of
their distance to the mole in question. The inverse square seems
reasonable as there can be very local distortions from image to image.

This is owing to the fact that skin stretches and compresses as the body
moves, and the body is not a perfect cylinder or plane. e.g. moles on a
bicep may move in a different direction to the rest of the arm if the
elbow angle is adjusted between images.

Also, automatically adjust the error bounds of the offset guessing.
Adjust it according to the weighted sum of the error of the surrounding
offsets, if they were to be replaced uniformly with the new guessed
offset.

Test Plan:
.. rotomap-edit and use 't' to relate moles on an upper arm ..
.. observe that non-uniform motion between images is handled better ..