Add 'mel-debug bench-relate' for testing relate
When making improvements to the routines for relating rotomaps to
eacheother, it would be useful to be able to easily see the global
impact of the change. This would help avoid local optimisations that are
actually global pessimisations.

Provide a utility that can operate on rotomaps that have already been
correctly related to eachother, noting where the routines for relating
provide incorrect answers.

An abbreviated example:

    $ mel-debug bench-relate *.jpg
    True negative: (00.jpg -> 01.jpg), (None -> c2b...)
    True positive: (00.jpg -> 01.jpg), (a44... -> a44...)
    False positive: (00.jpg -> 01.jpg), (d4c... -> 01b...)
    ...

Test Plan:

    # Run on all rotomaps, count the number of 'false' mappings and
    # 'true' mappings.
    $ find rotomaps -type d -links 2 |\
        xargs -I {} bash -c 'mel-debug bench-relate {}/*.jpg' |\
        cut -d' ' -f 1 | sort | uniq -c
    .. False mappings are roughly 10% of true mappings ..