mel-debug bench-relate: add '--reset-uuids' arg
When using 'bench-relate' to test, we don't only want to test that
spurious additional mappings are made. We also want to test that the
existing, canonical, mappings can be reproduced by the routines.

Provide a '--reset-uuids' option which will reset the uuids of a
specified number of moles, and thereby clear out any possible existing
mapping. It tries all the different combinations of resetting this
number of moles.

This will of course skew the output towards showing the error rate for
rotomaps with a large number of moles, as there will be more mappings
printed.

Test Plan:

Add a test for 'yield_reset_combinations()'.

Use the awesome GNU Parallel to process the much larger dataset
presented by the combinatorial expansion. Made a small donation via
https://my.fsf.org/donate, in the hope that it continues being
supported.

    # Test re-mapping moles on Right{Arm,Leg} rotomaps
    $ time find Right* -type d -links 2 | \
        parallel \
            'parallel echo mel-debug bench-relate --reset-uuids 5 \
                :::: <(ls {}/*.jpg) \
                ::::+ <(ls {}/*.jpg | tail -n +2)' | \
        parallel --bar | \
        cut -d' ' -f 1 | \
        sort | \
        uniq -c

    196980 False
    9951301 True

    real    2m38.754s
    user    18m14.591s
    sys     0m44.569s
    .. So about 2% falsehood apparently ..