mel-debug bench-relate: add '--reset-all-uiuds'
When comparing the performance of rotomap-relate routines with
'bench-relate', it's interesting to also observe what happens when all
the associations between moles are reset. This is the usual case when
creating a rotomap from scratch.

Provide an option for explicitly resetting all the moles. This would
have been possible previously by supplying a large number for
'--reset-uuids'.

Test Plan:

    # Test re-mapping moles on Right{Arm,Leg} rotomaps, reset all moles.
    $ find Right* -type d -links 2 | \
        parallel \
            'parallel echo mel-debug bench-relate --reset-all-uuids \
                :::: <(ls {}/*.jpg) \
                ::::+ <(ls {}/*.jpg | tail -n +2)' | \
        parallel --bar | \
        cut -d' ' -f 1 | \
        sort | \
        uniq -c

    915 False
    64 Flawed
    107 Flawless
    1446 True
    .. So curiosly, there are less errors when remapping all moles ..
    .. than when remapping up to 5 of them ..