mel.rotomap.relate: use _MAGIC_FIELD_ERROR value
After experimentation with 'mel-debug bench-relate --reset-uuids 1', it
turns out that pick_value_from_field() is performing much worse than
what came before. It turns out that this is due to the error bounds it
calculates.

Discovering things like this was the motivation behind creating
'bench-relate', so it's now possible to revisit assumptions with facts.

Instead of calculating the error bounds to use, simply pick a number
that works for a reasonable number of images and use that instead. The
performance is dramatically better.

In later work, we'll want to find a way of automatically determining the
best error value to use.

See the introduced comments for test results.

Out of curiosity I also removed the 'reflexive check' from
make_offset_field_theory(), to make sure it has an effect, it is better
with than without:
  64 Flawed
 268 Flawless