mel-debug relate-bench: fix flaw reporting
Change relate-bench to report only items that appear in the generated
theory as flaws. This means that only things that appear in the theory
are reported as False +ves or -ves.

The expected results shouldn't be reported as False +ves or -ves,
although they may be considered as 'missed' mappings.

Test Plan:
.. 'relate-bench' on an image with only known false -ves ..
.. See that before things were reported twice, now only false -ves ..