mel rotomap-edit: experimental lk_point_offsets()
Add an experimental new function for guessing the relationship between
images by using some of the optical flow routines available in OpenCV.

Use cv2.goodFeaturesToTrack() and cv2.calcOpticalFlowPyrLK() to try to
guess point_offsets between images.

Add debug rendering to make it easy to evaluate the guesses.

The guesses don't seem to be particularly good when inspecting visually.
The performance of 'bench-relate' is slightly worse with
'--iterate-relate'. The downside is possibly limited by the fact it's
only used for refining guesses, not the basis for a from-scratch guess.

Perhaps in later work we'll try different parameters to see if it can
work better. Maybe if we were working with video instead of photo sets
then the similarity would help give better results.

Test Plan:
.. Visually inspected with debug render, not great ..
.. Tested with 'mel-debug bench-relate', slightly worse ..