rotomap-edit: '3' for mole marking mode
Add a new mode that's tailored to the task of adding a marker where
there isn't one already.

The idea is to make it easy to see where there are already markers,
without worrying about obscuring the moles that they mark.

The task of removing erroneous markers, which are present where there
are no moles, is expected to be tackled separately.

Make the markers visually distinct from moles, so that it's easier to
focus on noticing only moles that haven't been marked.

Test Plan:
.. 'mel rotomap-edit' a rotomap, marking unmarked moles ..