rotomap-edit: 'a' in mole mark mode to accentuate
Reveal the moles that have been marked, whilst still showing markers.
This is good for verifying that markers are actually positioned on
moles.

Fade the parts of the image that have not been marked, to make it easier
for the eye to travel to where the marks are.

Test Plan:
.. 'mel rotomap-edit' a rotomap, removing moles in 'a' mode ..