rotomap-edit: 'image relate' mode
Introduce a mode specifically for relating the moles in one image to
another. The visualisation makes it easy to see the current
relationships between moles, as well as previewing automatic changes.

The actual moles are de-emphasised, as is the whole original image. This
makes it easier to focus on the relationships and to spot unlikely
mappings.

In later work we'll want to try to emphasise the moles as well as the
relationships.

Test Plan:
.. 'rotomap-edit' some rotomaps, relating to self and previous maps ..