rotomap-edit: in relate mode, r-click copies any
Instead of right-click only copying the 'from moles', also copy from the
'to moles'. This makes it much easier to copy moles when in the
target-mode of relate mode, as you only see the photograph of the 'to
moles'.

So you can right-click on the photograph of the mole you'd like to copy,
move to the next image, and then paste on the photograph of the mole
you'd like to relate it to.

Test Plan:
.. copy and paste to relate moles in target-mode ..