Fix rotomap.display to import rotomap.mask itself
This was previously undetected because when a submodule is imported
anywhere, it is automatically added to the namespace of the parent
package everywhere.

This means that if anything else first imports 'mel.rotomap.mask' and
then 'mel.rotomap.display' is used, no problems would be seen from this
missing import.

There are probably more of these mistakes lurking around.