Add 'mel rotomap-automask-svm' for trialling
Introduce a new subcommand for automasking. Use the new mel.lib.svm
classifier, to try to improve on the histogram-distance method.

This approach has the advantage of improving with extra training, which
isn't true of the histogram-distance method.

Test Plan:

Try different values of '--svm-c' and '--svm-gamma', to try to find
sensible defaults.

Visually compare results against those obtained by the
histogram-distance method, seems like it generally performs as good or
better.

It seems that cases where the performance is bad, it is generally
improved by providing extra training for those cases.