rotomap-{edit,automark}: no automole at mask edges
Reduce the number of false positives when automatically marking moles by
rejecting moles that are detected at the edges of a mask region. Those
edges are primarily where they seem to appear.

Use a magic number to configure the required distance for now. In later
work we can expose this to the CLI of 'mel rotomap-automark' and
friends.

Test Plan:
.. mel rotomap-automark, notice lots of false +ves before change ..
    .. notice far fewer false +ves after change ..
    .. notice about the same number of true +ves after change ..