rotomap-edit: move controls help to '--help'
Unclutter the command-line output by moving the help text instead to the
output of the '--help' argument.

This also moves towards the possibility of using stdout of this command
as the input to other things. Perhaps being able to replace mole-picker.

Additionally, it's much nicer not to have lots of print() commands. The
controls string needed extracting anyway.