Make 'mel' a minimal Python package
Now you can run 'pip3 install .' in a copy of the repository in order to
install mel. This is important for re-using mel internals in other
things. For example, working in Jupyter Notebooks.

Fix up tests to work on new location of package.

At this time it looks like the name 'mel' is not taken on PyPI.

Test Plan:
$ pip3 install --editable .
$ mel
.. shows command help ..
$ mel-debug
.. shows command help ..