MoleAcquirer: converge faster, speed up captures
Greatly increase the rate at which we converge on changing paramters in
the MoleAcquirer. This means that noisier input will be more tolerated,
and consequently that it will be faster to capture moles.

Test Plan:

Re-capture a mole where it seemed to be to fussy about the stability of
the captured mole.

Also tried with MoleAcquirer removed from 'mel micro-add' completely,
too many false +ves were captured in that case.