mel micro-add, no more rotation sensitivity
In the interest of greatly speeding up mole capture, remove rotation
sensitivity completely. It's easy enough to revert this commit later if
it actually causes problems.

It feels like getting the rotation straight contributes a lot of delay
during capture, but doesn't add a great amount of value.

Test Plan:

Re-capture a challenging mole before and after, there doesn't seem to be
a great deal lost in the new images and it felt signiificantly faster.