mel rotomap-montage-single: no outer border
Remove the outer border of the montage, it doesn't seem to be helpful in
any of the contexts I see it in.

Half the inner border to 25 pixels, the border seemed more thick than
necessary to distinguish the images.

Test Plan:
Re-generate one context image for micro capturing with updated command,
seems better.