mel rotomap-montage-single: add '--rot90'
Rotomap images aren't always immediately useful as context images for
self-imaging with a microscope, because of the change of point of view.

For example, if you are capturing an image of a mole on your forearm, if
the image has your elbow at the top of the image and your hand at the
bottom, you have to mentally re-orientate the image when looking down at
your actual forearm.

Reduce cognitive burden by providing a convenient '--rot90' argument
for rotating the images before they are montaged.

Test Plan:
Re-create a context image with the new parameter, observe that it's
easier to relate images to reality.