mel rotomap-montage-single: image write can fail
Note that cv2.imwrite() will return False on failure, it doesn't appear
to raise an exception.

Write a new helper function to raise an exception and use it instead
from rotomapmontagesingle.

This is particularly important when working with 'git annex', as images
will commonly be write-protected. It's important to fail visibly when
failing to overwrite these images.

Test Plan:
Try to use 'mel rotomap-montage-single' to overwrite a protected image,
see that after the changes it does fail as expected.