Refactor, use mel.lib.common.write_image() more
Given that cv2.imwrite() doesn't raise on failure, as perhaps was
expected, use the new mel.lib.common.write_image() instead of
cv2.imwrite().

Test Plan:
Hope that run_tests.sh will cover it.

Check that missing imports are picked up by the tests, unfortunately they
are not. It seems that a missing import error immediately on use would
be preferable to silent failure, so the lack of testing seems bearable.