mel micro-add: 'a' to abort, refactor
Change behaviour to accept only the 'a' key to abort capturing, this is
consistent with the outer loop, which also expects 'a' to abort.

Introduce a yield_frames_keys() function to deal with the fiddly bits of
video frame and key handling, this simplifies the capture loop a bit.

Test Plan:

Try all keyboard options of inner and outer capture loop, check return
codes.