Change 'mel list' to fail on ambiguous mole dirs
Fail when encountering mole directories that are ambiguous, e.g.
seemingly both a mole and a parent dir to a mole. The previous behaviour
was to treat the potential mole directory as a parent dir and ignore
it's mole-like files.

An example of a problematic directory:

    $ find .
    ./__id__
    ./__micro__
    ./2017.jpg
    ./AnotherMole/2017.jpg

Also, introduce a new package 'mel.micro' for sharing 'mel micro-*'
related routines.

Test Plan:

Use the updated 'mel list' on a large catalog of moles. Manually fixup
the cases complained about. Use sorting, assistance filters,
--only-no-micro.